requirements.txt
# Core ML/AI frameworks
torch>=2.0.0
torchaudio>=2.0.0

# Speech processing libraries
speechbrain>=0.5.0
pyannote.audio>=3.1.0
openai-whisper>=20231117

# Vector database and embeddings
pinecone>=3.0.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
googleapis-common-protos>=1.62.0

# Embedding models
cohere>=5.0.0

# Environment and configuration
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.24.0

# Audio processing dependencies
# librosa>=0.10.0  # Commented out due to Windows compatibility issues
soundfile>=0.12.0

# Additional dependencies for speech processing
scipy>=1.10.0
scikit-learn>=1.3.0

# Hugging Face transformers (required by pyannote.audio)
transformers>=4.30.0
datasets>=2.12.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Text processing and NLP
regex>=2023.0.0

# Windows-specific audio dependencies (alternative to librosa)
# ffmpeg-python>=0.2.0  # Uncomment if needed for audio processing

# Optional: CUDA support (uncomment if using GPU)
# torch-audio>=2.0.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# torch>=2.0.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
